+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Next
+
+## Features
+
+- Add support for `icrc21_canister_call_consent_message` to `@dfinity/ledger-icp`.
+
 # 2024.09.02-0830Z
 
 ## Overview